+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 ## 0.5.0 (2018-07-21)
 
+* Session cookie configuration
+
 * Actix/Actix-web 0.7 compatibility